Screen/FB/TopCanvas: add explicit cast to avoid narrowing warning

+++ src/Screen/FB/TopCanvas.cpp
@@ -451,7 +451,7 @@ TopCanvas::Flip()
       0, 0, buffer.width, buffer.height
     },
 
-    enable_dither ? WAVEFORM_MODE_A2 : WAVEFORM_MODE_AUTO,
+    uint32_t(enable_dither ? WAVEFORM_MODE_A2 : WAVEFORM_MODE_AUTO),
     UPDATE_MODE_FULL, // PARTIAL
     epd_update_marker,
     TEMP_USE_AMBIENT,
